did some formating also fixed contact in game.java Signed-off-by: DiaAWAY <sindrerh@hotmail.com>

+++ TDT4240_project/src/org/group20/sunstruck/Upgrade.java
@@ -11,10 +11,10 @@ public class Upgrade {
 	public static boolean upgrade(GameObject go, TYPE t, int amount) {
 		switch (t) {
 		case ARMOUR:
-			go.setHull((float) (go.getHull() + amount));
+			go.setHull((go.getHull() + amount));
 			break;
 		case SHIELD:
-			go.setShield((float) (go.getShield() + amount));
+			go.setShield((go.getShield() + amount));
 			break;
 		case SPEED:
 			go.setSpeed(go.getSpeed() + amount);
